Debt Record Printing
Debtor Listing now supports Printing of PDF. Should be working fine now!

How's the Google Maps thing btw? Probably going to be a bit rushy, since we have to come up with the poster, movie and the documentation too, IIRC. I'll have to start immediately on the algorithm soon, so GL!